requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
datamodels==0.8.2
exceptiongroup==1.0.4
fastapi==0.88.0
fastapi-responses==0.2.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
iniconfig==1.1.1
magic-filter==1.0.9
multidict==6.0.2
orjson==3.8.3
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.2
python-dotenv==0.21.0
pytz==2022.6
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
yarl==1.8.1
